+++ app/vmui/packages/vmui/src/components/Chart/BarHitsChart/BarHitsChart.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from "preact/compat";
+import { FC, useMemo, useState } from "preact/compat";
 import "./style.scss";
 import "uplot/dist/uPlot.min.css";
 import { AlignedData } from "uplot";
@@ -8,6 +8,8 @@ import { LogHits } from "../../../api/types";
 import { GraphOptions, GRAPH_STYLES } from "./types";
 import BarHitsOptions from "./BarHitsOptions/BarHitsOptions";
 import BarHitsPlot from "./BarHitsPlot/BarHitsPlot";
+import { calculateTotalHits } from "../../../utils/logs";
+import { LOGS_GROUP_BY, LOGS_LIMIT_HITS } from "../../../constants/logs";
 
 interface Props {
   logHits: LogHits[];
@@ -24,6 +26,7 @@ const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onAp
     hideChart: false,
   });
 
+  const totalHits = useMemo(() => calculateTotalHits(logHits), [logHits]);
 
   return (
     <div
@@ -32,9 +35,18 @@ const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onAp
         "vm-bar-hits-chart__wrapper_hidden": graphOptions.hideChart
       })}
     >
+      <div className="vm-bar-hits-chart-header">
+        <div className="vm-bar-hits-chart-header-info">
+          <p>Top <b>{LOGS_LIMIT_HITS}</b> hits grouped by <b>{LOGS_GROUP_BY}</b></p>
+          |
+          <p>Total: <b>{totalHits.toLocaleString("en-US")}</b> hits</p>
+        </div>
+        <BarHitsOptions onChange={setGraphOptions}/>
+      </div>
       {!graphOptions.hideChart && (
         <BarHitsPlot
           logHits={logHits}
+          totalHits={totalHits}
           data={_data}
           period={period}
           setPeriod={setPeriod}
@@ -42,7 +54,6 @@ const BarHitsChart: FC<Props> = ({ logHits, data: _data, period, setPeriod, onAp
           graphOptions={graphOptions}
         />
       )}
-      <BarHitsOptions onChange={setGraphOptions}/>
     </div>
   );
 };