+++ app/vmui/packages/vmui/src/components/Chart/BarHitsChart/LegendHitsMenu/LegendHitsMenu.tsx
@@ -1,4 +1,5 @@
 import { FC } from "preact/compat";
+import { Series } from "uplot";
 import "./style.scss";
 import { LegendLogHits } from "../../../../api/types";
 import LegendHitsMenuStats from "./LegendHitsMenuStats";
@@ -6,25 +7,34 @@ import LegendHitsMenuBase from "./LegendHitsMenuBase";
 import LegendHitsMenuRow from "./LegendHitsMenuRow";
 import LegendHitsMenuFields from "./LegendHitsMenuFields";
 import { LOGS_LIMIT_HITS } from "../../../../constants/logs";
+import LegendHitsMenuVisibility from "./LegendHitsMenuVisibility";
 
 const otherDescription = `aggregated results for fields not in the top ${LOGS_LIMIT_HITS}`;
 
 interface Props {
   legend: LegendLogHits;
   fields: string[];
+  series: Series[];
   onApplyFilter: (value: string) => void;
+  onRedrawGraph: () => void;
   onClose: () => void;
 }
 
-const LegendHitsMenu: FC<Props> = ({ legend, fields, onApplyFilter, onClose }) => {
+const LegendHitsMenu: FC<Props> = ({ legend, fields, series, onApplyFilter, onRedrawGraph, onClose }) => {
   return (
     <div className="vm-legend-hits-menu">
-      <div className="vm-legend-hits-menu-section">
-        <LegendHitsMenuRow
-          className="vm-legend-hits-menu-row_info"
-          title={legend.isOther ? otherDescription : legend.label}
-        />
-      </div>
+      {legend.isOther && (
+        <div className="vm-legend-hits-menu-section vm-legend-hits-menu-section_info">
+          <LegendHitsMenuRow title={otherDescription}/>
+        </div>
+      )}
+
+      <LegendHitsMenuVisibility
+        legend={legend}
+        series={series}
+        onRedrawGraph={onRedrawGraph}
+        onClose={onClose}
+      />
 
       {!legend.isOther && (
         <LegendHitsMenuBase